requirements.txt
flask
PyMuPDF
google-generativeai
fpdf
stripe